more externs

+++ include/m4a.h
@@ -13,7 +13,7 @@ void m4aSongNumStartOrChange(u16 n);
 void m4aSongNumStop(u16 n);
 void m4aMPlayAllStop(void);
 void m4aMPlayContinue(struct MusicPlayerInfo *mplayInfo);
-void m4aMPlayFadeOut(struct MusicPlayerInfo *mplayInfo, u32 speed);
+void m4aMPlayFadeOut(struct MusicPlayerInfo *mplayInfo, u16 speed);
 void m4aMPlayFadeOutTemporarily(struct MusicPlayerInfo *mplayInfo, u16 speed);
 void m4aMPlayFadeIn(struct MusicPlayerInfo *mplayInfo, u16 speed);
 void m4aMPlayImmInit(struct MusicPlayerInfo *mplayInfo);
@@ -21,7 +21,7 @@ void m4aMPlayStop(struct MusicPlayerInfo *mplayInfo);
 void m4aMPlayVolumeControl(struct MusicPlayerInfo *mplayInfo, u16 trackBits, u16 volume);
 void m4aSoundVSync(void);
 
-extern struct MusicPlayerInfo gMPlayInfo_BGM; // BGM??
+extern struct MusicPlayerInfo gMPlayInfo_BGM;
 extern struct MusicPlayerInfo gMPlayInfo_Fanfare;
 extern struct MusicPlayerInfo gMPlayInfo_SE1;
 extern struct MusicPlayerInfo gMPlayInfo_SE2;
